requirements.txt
google-generativeai>=0.3.1
google-adk>=0.4.0
streamlit>=1.31.0
python-dotenv>=1.0.0
pytest>=7.4.3 